add f32tos and Atob function to conv package
