+++ internal/conv/conv.go
@@ -20,11 +20,30 @@ import (
 	"unsafe"
 )
 
-// Btoa converts from bytes to type string
+// Btoa converts from byte slice to string.
 func Btoa(b []byte) string {
 	header := (*reflect.SliceHeader)(unsafe.Pointer(&b))
 	return *(*string)(unsafe.Pointer(&reflect.StringHeader{
 		Data: header.Data,
 		Len:  header.Len,
 	}))
+}
+
+// Atobs converts from string to byte slice.
+func Atob(s string) []byte {
+	return *(*[]byte)(unsafe.Pointer(&reflect.SliceHeader{
+		Data: (*(*reflect.StringHeader)(unsafe.Pointer(&s))).Data,
+		Len:  len(s),
+		Cap:  len(s),
+	}))
+}
+
+// F32stos converts from float32 slice to type string.
+func F32stos(fs []float32) string {
+	lf := 4 * len(fs)
+	buf := (*(*[1]byte)(unsafe.Pointer(&(fs[0]))))[:]
+	addr := unsafe.Pointer(&buf)
+	(*(*int)(unsafe.Pointer(uintptr(addr) + uintptr(8)))) = lf
+	(*(*int)(unsafe.Pointer(uintptr(addr) + uintptr(16)))) = lf
+	return *(*string)(unsafe.Pointer(&buf))
 }
