+++ internal/conv/conv_test.go
@@ -23,56 +23,7 @@ import (
 	"github.com/vdaas/vald/internal/test/goleak"
 )
 
-// NOTE: generated by https://json-generator.com/
-var testJson = `
-[
-  {
-    "_id": "6228304e91ff005521f055a0",
-    "index": 0,
-    "guid": "32a05fca-9759-41fa-882c-a878f670d3f5",
-    "isActive": false,
-    "balance": "$1,943.23",
-    "picture": "http://placehold.it/32x32",
-    "age": 33,
-    "eyeColor": "green",
-    "name": "Goodman Fitzgerald",
-    "gender": "male",
-    "company": "ORBIXTAR",
-    "email": "goodmanfitzgerald@orbixtar.com",
-    "phone": "+1 (808) 547-2333",
-    "address": "750 Sunnyside Avenue, Maybell, Vermont, 356",
-    "about": "In esse nisi quis nisi cupidatat. Qui esse magna consequat amet mollit adipisicing adipisicing proident. Consequat consequat quis velit non proident. Consequat commodo magna sit do labore. Sunt reprehenderit do id deserunt voluptate velit ullamco.\r\n",
-    "registered": "2018-10-17T03:11:06 -09:00",
-    "latitude": 4.696977,
-    "longitude": -90.128225,
-    "tags": [
-      "non",
-      "voluptate",
-      "excepteur",
-      "est",
-      "tempor",
-      "id",
-      "nostrud"
-    ],
-    "friends": [
-      {
-        "id": 0,
-        "name": "Marcy Sheppard"
-      },
-      {
-        "id": 1,
-        "name": "Santos Donaldson"
-      },
-      {
-        "id": 2,
-        "name": "Shari Orr"
-      }
-    ],
-    "greeting": "Hello, Goodman Fitzgerald! You have 7 unread messages.",
-    "favoriteFruit": "banana"
-  }
-]
-`
+var testData = "!\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz"
 
 func TestBtoa(t *testing.T) {
 	type args struct {
@@ -115,12 +66,12 @@ func TestBtoa(t *testing.T) {
 			},
 		},
 		{
-			name: "returns json string when bytes is json data",
+			name: "returns string when bytes is not empty",
 			args: args{
-				b: []byte(testJson),
+				b: []byte(testData),
 			},
 			want: want{
-				want: testJson,
+				want: testData,
 			},
 		},
 	}
@@ -145,6 +96,145 @@ func TestBtoa(t *testing.T) {
 			if err := checkFunc(test.want, got); err != nil {
 				tt.Errorf("error = %v", err)
 			}
+		})
+	}
+}
+
+func TestAtob(t *testing.T) {
+	type args struct {
+		s string
+	}
+	type want struct {
+		want []byte
+	}
+	type test struct {
+		name       string
+		args       args
+		want       want
+		checkFunc  func(want, []byte) error
+		beforeFunc func(args)
+		afterFunc  func(args)
+	}
+	defaultCheckFunc := func(w want, got []byte) error {
+		if !reflect.DeepEqual(got, w.want) {
+			return errors.Errorf("got: \"%#v\",\n\t\t\t\twant: \"%#v\"", got, w.want)
+		}
+		return nil
+	}
+	tests := []test{
+		{
+			name: "returns empty when str is empty",
+			args: args{
+				s: "",
+			},
+			want: want{},
+		},
+		{
+			name: "returns empty when s is not empty",
+			args: args{
+				s: testData,
+			},
+			want: want{
+				[]byte(testData),
+			},
+		},
+	}
+
+	for _, tc := range tests {
+		test := tc
+		t.Run(test.name, func(tt *testing.T) {
+			tt.Parallel()
+			defer goleak.VerifyNone(tt, goleak.IgnoreCurrent())
+			if test.beforeFunc != nil {
+				test.beforeFunc(test.args)
+			}
+			if test.afterFunc != nil {
+				defer test.afterFunc(test.args)
+			}
+			checkFunc := test.checkFunc
+			if test.checkFunc == nil {
+				checkFunc = defaultCheckFunc
+			}
+
+			got := Atob(test.args.s)
+			if err := checkFunc(test.want, got); err != nil {
+				tt.Errorf("error = %v", err)
+			}
+
+		})
+	}
+}
+
+func TestF32stos(t *testing.T) {
+	type args struct {
+		fs []float32
+	}
+	type want struct {
+		want string
+	}
+	type test struct {
+		name       string
+		args       args
+		want       want
+		checkFunc  func(want, string) error
+		beforeFunc func(args)
+		afterFunc  func(args)
+	}
+	defaultCheckFunc := func(w want, got string) error {
+		if !reflect.DeepEqual(got, w.want) {
+			return errors.Errorf("got: \"%#v\",\n\t\t\t\twant: \"%#v\"", got, w.want)
+		}
+		return nil
+	}
+	tests := []test{
+		// TODO test cases
+		/*
+		   {
+		       name: "test_case_1",
+		       args: args {
+		           fs: nil,
+		       },
+		       want: want{},
+		       checkFunc: defaultCheckFunc,
+		   },
+		*/
+
+		// TODO test cases
+		/*
+		   func() test {
+		       return test {
+		           name: "test_case_2",
+		           args: args {
+		           fs: nil,
+		           },
+		           want: want{},
+		           checkFunc: defaultCheckFunc,
+		       }
+		   }(),
+		*/
+	}
+
+	for _, tc := range tests {
+		test := tc
+		t.Run(test.name, func(tt *testing.T) {
+			tt.Parallel()
+			defer goleak.VerifyNone(tt, goleak.IgnoreCurrent())
+			if test.beforeFunc != nil {
+				test.beforeFunc(test.args)
+			}
+			if test.afterFunc != nil {
+				defer test.afterFunc(test.args)
+			}
+			checkFunc := test.checkFunc
+			if test.checkFunc == nil {
+				checkFunc = defaultCheckFunc
+			}
+
+			got := F32stos(test.args.fs)
+			if err := checkFunc(test.want, got); err != nil {
+				tt.Errorf("error = %v", err)
+			}
+
 		})
 	}
 }
